requirements.txt
torch
torchvision
lightning
hydra-core
rich
tensorboard
